+++ CHANGELOG.md
@@ -9,8 +9,9 @@
 
 ### Fixed
 
-- Don't panic if the collection file is invalid on first startup
+- Don't panic if the collection file is invalid on first startup [#34](https://github.com/LucasPickering/slumber/issues/34)
   - The TUI will now show an empty screen, and watch the collection file for changes
+- Fix long status code reasons getting cut off in response header [#40](https://github.com/LucasPickering/slumber/issues/40)
 
 ## [0.7.0] - 2023-11-16